Added the rest of the HTML
I added the newsletter and styled it
I added the footer and styled it
I fixed the header as it was misplaced. I removed the JS function and just made it fixed
I added the timer in the hero image. I added a JS function and made it work as intended to the end of the day