Added zoom effect
I added a jQuery plugin called Magnify JS for the zoom effect on the hero image. I decided to go with hover zoom effect.
I made it work at first but then it stopped working. I am looking for a solution. The magnifying glass appears but instead of showing a zoomed image it shows error message