Responsiveness
I added some responsiveness to the page.
I made sure that everything is visible and looks as intended.
I had a lot of issues with the image slider as it is far too difficult to style it.
I added a hamburger menu on small screens and added a function to open/close it. Also added some conditions to make sure it gets closed if the user resizes the page.
Made some changes to the product details part to make it flex-direction: column and that way it will appear under the image slider.
Also added some other styles to make sure the rest of the page is responsive.
Had some issues with the footer too. It has 4 different divs with all the links. I decided to decrease the font-size for iPad kind of screen and then changed the flex to column in order to show each one on a separate line.
Also added some validation and JavaScript to the newsletter form in order to make it submit. I added an alert on submission